+++ test/Service/AuthorizationServiceTest.php
@@ -24,8 +24,8 @@ use PHPUnit\Framework\TestCase;
 use ZfcRbac\Assertion\AssertionPluginManager;
 use ZfcRbac\Exception\InvalidArgumentException;
 use ZfcRbac\Identity\IdentityInterface;
-use ZfcRbac\Rbac\Rbac;
-use ZfcRbac\Rbac\Role\RoleInterface;
+use ZfcRbac\Rbac;
+use ZfcRbac\Role\RoleInterface;
 use ZfcRbac\Service\AuthorizationService;
 use ZfcRbac\Service\RoleService;
 use ZfcRbac\Service\RoleServiceInterface;
